이메일 서비스 추가

+++ intime/src/main/java/com/topcom/intime/service/GoogleService.java
@@ -36,7 +36,7 @@ public class GoogleService {
 
     public LocationDto getDatafromApi(int useridx, String address){
         if(!userRepository.existsById(useridx)){
-            throw new ResourceNotFoundException("User", "useridx", (long) useridx);
+            throw new ResourceNotFoundException("User", "useridx", useridx);
         }
         try{
             LocationDto locationDto=new LocationDto();
